requirements.txt
## The following requirements were added by pip freeze:
asgiref==3.8.1
Django==5.2.1
gitdb==4.0.12
GitPython==3.1.41
isoweek==1.3.3
python-dateutil==2.9.0.post0
screen==1.0.1
setuptools==75.6.0
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
TimeConvert==3.0.13
tzlocal==5.3.1
xlwt==1.3.0
djangorestframework==3.16.0
shortuuid==1.0.13
## The following requirements were added by pip freeze:
pillow==11.2.1